Add files via upload
Multiplexer is a combinational which transmit large number of information over a single line. The term "multiplex " means "many into one". A multiplexer is also called data selector since it selects one of many inputs and steers the information to the output. Multiplexer is designed in such a way with n input lines ,m select lines and one output line. If the no. of n input lines is equal to 2^m then m select lines are required to select one of n input lines. Multiplexer is a universal combinational circuit. I have designed 4x1 mux.