Add files via upload
Decoder is a combinational logic circuit that converts an n bit binary input code into 2^n output lines ,such that output lines will be activated for only one of the possible combinations of inputs. Decoding is necessary in applications such as data demutiplexing ,digital display, digital to analog converters and memory addressing. In a decoder the no. of output is greater than no. of inputs. It is important that  if no. of inputs is equal to no. of output then it is called converters ex BCD to excess3,Binary to gray etc. Designing 2x4 decoder using system verilog.